+++ esopie/model.py
@@ -11,7 +11,6 @@ from queue import Queue
 from PySide2.QtCore import Signal, Qt
 from PySide2.QtCore import QSize, Qt, Signal, QSettings
 
-from esopie.eso_file_header import FileHeader
 from esopie.utils.utils import generate_ids, get_str_identifier
 from esopie.utils.process_utils import (create_pool, kill_child_processes,
                                         load_file, wait_for_results)
@@ -20,7 +19,7 @@ from esopie.threads import (EsoFileWatcher, GuiMonitor, ResultsFetcher,
 
 
 class AppModel:
-    file_loaded = Signal(str, FileHeader, FileHeader)
+    fileLoaded = Signal(str, str)
 
     def __init__(self):
         settings = QSettings()
@@ -28,6 +27,7 @@ class AppModel:
         # ~~~~ Intermediate settings ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
         self.selected_variables = None
         self.all_files_selected = False
+        self.totals_selected = False
         self.units_settings = {
             "energy_units": settings.value("Toolbar/energyUnits", "kWh"),
             "power_units": settings.value("Toolbar/powerUnits", "kW"),
@@ -41,7 +41,8 @@ class AppModel:
         self.progress_queue = self.manager.Queue()
 
         # ~~~~ Database ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-        self.database = {}
+        self.database = {"standard": {},
+                         "totals": {}}
 
         # ~~~~ Process executor ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
         self.pool = create_pool()
@@ -59,11 +60,12 @@ class AppModel:
 
     def tear_down(self):
         self.watcher.terminate()
+        self.monitor.terminate()
         self.manager.shutdown()
 
         kill_child_processes(os.getpid())
 
-    def load_files(self, eso_file_paths):
+    def process_eso_files(self, eso_file_paths):
         """ Start eso file processing. """
         used_ids = self.get_used_ids_from_db()
         ids = generate_ids(used_ids, n=len(eso_file_paths))
@@ -99,19 +101,143 @@ class AppModel:
         std_file.rename(name)
         tot_file.rename(f"{name} - totals")
 
-        # copy header dicts as view and file should be independent
-        std_header_dct = deepcopy(std_file.header_dct)
-        tot_header_dct = deepcopy(tot_file.header_dct)
-
-        std_header = FileHeader(std_id, std_header_dct)
-        tot_header = FileHeader(tot_id, tot_header_dct)
-
         file_set = {std_id: std_file,
                     tot_id: tot_file}
 
         self.add_set_to_db(id_, file_set)
 
-        self.file_loaded.emit(id_, name, std_header, tot_header)
+        self.fileLoaded.emit(id_, name)
+
+    def remove_vars(self):
+        """ Remove variables from a file. """
+        variables = self.get_current_request()
+
+        if not variables:
+            return
+
+        all_ = self.toolbar.all_files_requested()
+        nm = self.tab_wgt.tabText(self.tab_wgt.currentIndex())
+
+        files = "all files" if all_ else f"file '{nm}'"
+        text = f"Delete following variables from {files}: "
+
+        inf_text = "\n".join([" | ".join(var[1:3]) for var in variables])
+
+        dialog = ConfirmationDialog(self, text, det_text=inf_text)
+        res = dialog.exec_()
+
+        if res == 0:
+            return
+
+        self.apply_async(self.dump_vars, variables, remove=True)
+        self.rebuild_view()
+
+    def hide_vars(self):
+        """ Temporarily hide variables. """
+        variables = self.get_current_request()
+        self.apply_async(self.dump_vars, variables)
+
+        # allow showing variables again
+        self.show_hidden_act.setEnabled(True)
+        self.toolbar.hide_btn.setEnabled(True)
+
+        self.rebuild_view()
+
+    def add_var(self, aggr_func):
+        """ Create a new variable using given aggr function. """
+        variables = self.get_current_request()
+
+        # retrieve variable name from ui
+        msg = "Enter details of the new variable: "
+        res = self.get_new_name(variables, msg=msg)
+
+        if res:
+            var_nm, key_nm = res
+            var = self.apply_async(self.aggr_vars, var_nm, key_nm,
+                                   variables, aggr_func)
+
+            self.selected = [var]
+            self.rebuild_view()
+            self.current_view_wgt.scroll_to(var)
+
+    def aggr_vars(self, view, var_nm, key_nm, variables, func):
+        """ Add a new aggreagated variable to the file. """
+        file_id = view.get_file_id()
+
+        # files are always returned as list
+        file = self.get_files_from_db(file_id)[0]
+
+        res = file.aggregate_variables(variables, func,
+                                       key_nm=key_nm,
+                                       var_nm=var_nm,
+                                       part_match=False)
+        if res:
+            var_id, var = res
+            view.set_next_update_forced()
+
+            return var
+
+    def dump_vars(self, view, variables, remove=False):
+        """ Hide or remove selected variables. """
+        file_id = view.get_file_id()
+        file = self.get_files_from_db(file_id)[0]
+
+        groups = file.find_pairs(variables)
+
+        if not groups:
+            return
+
+        if remove:
+            view.remove_header_variables(groups)
+            file.remove_outputs(variables)
+        else:
+            view.hide_header_variables(groups)
+
+        view.set_next_update_forced()
+
+    def apply_async(self, func, *args, **kwargs):
+        """ A wrapper to apply functions to current views. """
+        # apply function to the current widget
+        view = self.current_view_wgt
+        val = func(view, *args, **kwargs)
+
+        # apply function to all other widgets asynchronously
+        others = self.all_other_view_wgts
+        if others:
+            w = IterWorker(func, others, *args, **kwargs)
+            self.thread_pool.start(w)
+
+        return val
+
+    def show_hidden_vars(self):
+        """ Show previously hidden variables. """
+        for view in self.current_view_wgts:
+            view.show_hidden_header_variables()
+            view.set_next_update_forced()
+
+        self.show_hidden_act.setEnabled(False)
+        self.toolbar.hide_btn.setEnabled(False)
+
+        self.rebuild_view()
+
+    def remove_hidden_vars(self):
+        """ Remove hidden variables. """
+        for view in self.current_view_wgts:
+            view.remove_hidden_header_variables()
+
+    def rename_var(self, view, var_nm, key_nm, variable):
+        """ Rename given 'Variable'. """
+        file_id = view.get_file_id()
+        file = self.get_files_from_db(file_id)[0]
+
+        res = file.rename_variable(variable, var_nm, key_nm)
+        if res:
+            var_id, var = res
+            # add variable will replace current variable
+            view.add_header_variable(var_id, var)
+            view.set_next_update_forced()
+
+            return var
 
     def get_results(self, callback=None, **kwargs):
         """ Get output values for given variables. """
